requirements.txt
albumentations==1.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
bleach==5.0.0
brotlipy==0.7.0
certifi==2021.5.30
conda==4.10.1
conda-build==3.21.4
debugpy==1.6.0
defusedxml==0.7.1
dnspython==2.1.0
docopt==0.6.2
entrypoints==0.4
fastjsonschema==2.15.3
imageio==2.27.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==7.32.0
jedi==0.17.0
joblib==1.2.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab==2.2.5
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.0
mkl-service==2.3.0
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.6.3
notebook==6.4.11
olefile==0.46
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==21.3
pandocfilters==1.5.0
Pillow==9.4.0
pipreqs==0.4.13
pkginfo==1.7.0
prometheus-client==0.14.1
pycosat==0.6.3
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-etcd==0.4.5
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==22.3.0
qudida==0.0.4
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
terminado==0.13.3
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
torch==1.9.0
torchelastic==0.2.0
torchtext==0.10.0
torchvision==0.10.0
tornado==6.1
traitlets==5.1.1
typing-extensions==4.5.0
webencodings==0.5.1
yarg==0.1.9
zipp==3.8.0
